requirements.txt
streamlit 
langchain-openai 
langchain-community 
pillow 
python-dotenv
duckduckgo-search
wikipedia